fix: update structure
