+++ src/components/Dashboard/Explore/WorldMap/NodesMapView.js
@@ -53,30 +53,6 @@ export default function NodesMapView({ nodes = [] }) {
   //   }
   // }, [nodes]);
 
-  // nodes = [
-  //   {
-  //     latitude: 43,
-  //     longitude: -79,
-  //     status: true,
-  //     name: 'Rong',
-  //     created: '9-21-2022 12:00',
-  //     remark: 'mock hive node 1',
-  //     ip: '10.10.10.10',
-  //     url: 'testnet-hive1.trinity-tech.io',
-  //     owner_did: 'did:elastos:ijvDXRSCwFF9uMzpdLMZJW2e8i6xSjAkxQ'
-  //   },
-  //   {
-  //     latitude: 20,
-  //     longitude: 73,
-  //     status: false,
-  //     name: 'Frost',
-  //     created: '9-23-2022 12:00',
-  //     remark: 'mock hive node 2',
-  //     ip: '10.10.10.10',
-  //     url: 'testnet-hive2.trinity-tech.io',
-  //     owner_did: 'did:elastos:ijvDXRSCwFF9uMzpdLMZJW2e8i6xSjAkxQ'
-  //   }
-  // ];
   const isLoaded = nodes.length > 0 && Object.keys(nodes[0]).length > 0;
   const handleAccess = () => {};
